+++ Cloner/Cloner.php
@@ -11,7 +11,7 @@
 namespace Darvin\Utils\Cloner;
 
 use Darvin\Utils\Event\CloneEvent;
-use Darvin\Utils\Event\Events;
+use Darvin\Utils\Event\CloneEvents;
 use Darvin\Utils\Mapping\Annotation\Clonable\Clonable;
 use Darvin\Utils\Mapping\MappingException;
 use Darvin\Utils\Mapping\MetadataFactoryInterface;
@@ -135,7 +135,7 @@ class Cloner implements ClonerInterface
         }
 
         $event = new CloneEvent($object, $clone);
-        $this->eventDispatcher->dispatch(Events::POST_CLONE, $event);
+        $this->eventDispatcher->dispatch(CloneEvents::CREATED, $event);
 
         return $event->getClone();
     }