+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 6.5.4: Add "inverse" option to new object flag annotation.
 
 6.5.6: Add importable interface.
+
+6.6.0: Add "format_price" Twig filter.